requirements.txt
undetected-chromedriver
selenium
beautifulsoup4
paho-mqtt
